requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
discord.py==1.6.0
Flask==1.1.2
gevent==21.1.2
greenlet==1.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==5.1.0
Pillow==8.1.0
python-dotenv==0.15.0
requests==2.25.1
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
yarl==1.6.3
zope.event==4.5.0
zope.interface==5.2.0
